Stateful ROS on Core x 1.01

``` ini

BenchmarkDotNet=v0.11.0, OS=Windows 10.0.17730
Intel Core i7-4720HQ CPU 2.60GHz (Haswell), 1 CPU, 8 logical and 4 physical cores
.NET Core SDK=2.1.300
  [Host] : .NET Core 2.1.2 (CoreCLR 4.6.26628.05, CoreFX 4.6.26629.01), 64bit RyuJIT
  Clr    : .NET Framework 4.7.2 (CLR 4.0.30319.42000), 64bit RyuJIT-v4.7.3062.0
  Core   : .NET Core 2.1.2 (CoreCLR 4.6.26628.05, CoreFX 4.6.26629.01), 64bit RyuJIT

```
|         Method |  Job | Runtime |     Mean |     Error |    StdDev | Scaled | ScaledSD |    Gen 0 |   Gen 1 | Allocated |
|--------------- |----- |-------- |---------:|----------:|----------:|-------:|---------:|---------:|--------:|----------:|
|   MemoryStream |  Clr |     Clr | 2.178 ms | 0.0086 ms | 0.0076 ms |   1.00 |     0.00 | 148.4375 | 46.8750 | 734.82 KB |
|            ROS |  Clr |     Clr | 3.571 ms | 0.0105 ms | 0.0093 ms |   1.64 |     0.01 | 152.3438 | 50.7813 | 734.84 KB |
|   Stateful ROS |  Clr |     Clr | 2.761 ms | 0.0128 ms | 0.0120 ms |   1.27 |     0.01 | 144.5313 | 50.7813 | 734.74 KB |
|                |      |         |          |           |           |        |          |          |         |           |
|   MemoryStream | Core |    Core | 2.320 ms | 0.0142 ms | 0.0133 ms |   1.00 |     0.00 | 148.4375 | 46.8750 |  734.8 KB |
|            ROS | Core |    Core | 3.400 ms | 0.0435 ms | 0.0407 ms |   1.47 |     0.02 | 152.3438 | 50.7813 | 734.81 KB |
|   Stateful ROS | Core |    Core | 2.334 ms | 0.0402 ms | 0.0335 ms |   1.01 |     0.01 | 144.5313 | 50.7813 | 734.71 KB |