@benaadams 2018-08-06
Manually inline Peek to reduce span copies, Jit doesn't remove the copies after inlining

Don't pass mutable struct via `in` then call a property on it as that will cause a copy; so always pass by `ref`
- had to inline a method for the stackalloc version as compiler got unhappy

Only first arguments get passed by register, no point in wasting on Span which gets passed via stack